Chore: #7- 게임 로직 분리
- 게임 관련 로직 분리
